+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Types of changes:
 
 - `test_load_hf_components:test_get_model_hf()` to assert model type
 - `coverage` for reporting of functions covered with `pytest`
+- `tests/conftest.py` for pytest fixtures
 
 ### Changed